# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para consulta de codigos postales",
        "description": "Busca de codigos postales, información de estados, municipios y sus asentamientos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://codigo-postal.test/api"
        }
    ],
    "paths": {
...
        "/cp/{cp}": {
            "get": {
                "tags": [
                    "cp"
                ],
                "summary": "Obtener asentamientos por codigo postal",
                "operationId": "dea6839a2591b25c535dabbc67b79d3b",
                "parameters": [
                    {
                        "name": "cp",
                        "in": "path",
                        "description": "Codigo postal para buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "cp": {
                                                        "type": "number",
                                                        "example": 97825
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Mukuiche"
                                                    },
                                                    "tipo_asentamiento": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "tipo": {
                                                                "type": "string",
                                                                "example": "Pueblo"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "ciudad": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "null"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "municipio": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Abalá"
                                                            },
                                                            "estado": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "number",
                                                                        "example": "1"
                                                                    },
                                                                    "nombre": {
                                                                        "type": "string",
                                                                        "example": "Yucatán"
                                                                    },
                                                                    "abrev": {
                                                                        "type": "string",
                                                                        "example": "Yuc."
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso Suburb no encontrado, puede que el id no exista en la base de datos"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}